Fix errors with types
