+++ packages/components/src/components/buttons/index.ts
@@ -21,7 +21,6 @@ export * from './Action'
 export * from './Button'
 export * from './PrimaryButton'
 export * from './SecondaryButton'
-export * from './ListItemAction'
 export * from './ExpansionButton'
 export * from './LinkButton'
 export * from './TertiaryButton'
