+++ packages/components/src/components/interface/index.ts
@@ -20,8 +20,6 @@ export * from './DataTable/Pagination'
 export * from './SelectGroup'
 export * from './ListViewSimplified/ListViewSimplified'
 export * from './DataTable/SortAndFilter'
-export * from './ListItem'
-export * from './ListItemExpansion'
 export * from './DocumentViewer'
 export * from './SectionDrawer'
 export * from './Divider'
